requirements.txt
# Flask and extensions
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Werkzeug==2.3.7
Flask-Mail==0.9.1
Flask-WTF==1.2.1

# Database
SQLAlchemy==2.0.20

# Security
bcrypt==4.0.1
python-dotenv==1.0.0
email-validator==2.1.0

# Machine Learning
tensorflow==2.13.0
keras==2.13.1
scikit-learn==1.3.0
joblib==1.3.2

# Data Processing
pandas==2.1.0
numpy==1.24.3
python-dateutil==2.8.2
pytz==2023.3

# Data Visualization (backend)
matplotlib==3.7.2
seaborn==0.12.2
reportlab==4.0.4
matplotlib==3.7.1

# Development & Testing
pytest==7.4.2
pytest-cov==4.1.0